fix naive beefy prover
