+++ modules/trees/mmr/primitives/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mmr-primitives"
-version = "1.15.1"
+version = "1.15.2"
 edition = "2021"
 authors = ["Polytope Labs <hello@polytope.technology>"]
 license = "Apache-2.0"
